google-vertex-ai-workbench/requirements.txt
google-cloud-aiplatform
python-dotenv
jupyterlab
ipykernel
transformers
torch
